# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/encounters": {
      "get": {
        "operationId": "EncountersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/EncounterResponseDto"
                          }
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "encounters"
        ]
      },
      "post": {
        "operationId": "EncountersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEncounterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/EncounterResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "encounters"
        ]
      }
    },
...
    "/api/v1/encounters/{encounterId}/ai-assessments/{aiAssessmentId}/reviews": {
      "post": {
        "operationId": "EncounterDoctorDecisionController_reviewAssessment",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "aiAssessmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewAssessmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DoctorReviewResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "doctor-decision"
        ]
      }
    },
    "/api/v1/encounters/{encounterId}/ai-assessments/{aiAssessmentId}/review": {
      "post": {
        "operationId": "EncounterDoctorDecisionController_reviewAssessmentAlias",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "aiAssessmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewAssessmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "doctor-decision"
        ]
      }
    },
    "/api/v1/encounters/{encounterId}/reviews": {
      "get": {
        "operationId": "EncounterDoctorDecisionController_listReviews",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/DoctorReviewResponseDto"
                          }
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "doctor-decision"
        ]
      }
    },
    "/api/v1/encounters/{encounterId}/diagnoses": {
      "post": {
        "operationId": "EncounterDoctorDecisionController_recordDiagnosis",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordDiagnosisRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DoctorDiagnosisResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "doctor-decision"
        ]
      },
      "get": {
        "operationId": "EncounterDoctorDecisionController_listDiagnoses",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/DoctorDiagnosisResponseDto"
                          }
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "doctor-decision"
        ]
      }
    },
    "/api/v1/encounters/{encounterId}/clinical-plan": {
      "post": {
        "operationId": "EncounterDoctorDecisionController_approveClinicalPlan",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveClinicalPlanRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ClinicalPlanResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "doctor-decision"
        ]
      },
      "get": {
        "operationId": "EncounterDoctorDecisionController_getClinicalPlan",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ClinicalPlanResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "doctor-decision"
        ]
      }
    },
    "/api/v1/diagnoses/{diagnosisId}/revisions": {
      "post": {
        "operationId": "DiagnosesController_reviseDiagnosis",
        "parameters": [
          {
            "name": "diagnosisId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseDiagnosisRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DoctorDiagnosisResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "doctor-decision"
        ]
      }
    },
    "/api/v1/diagnoses/{diagnosisId}/revise": {
      "post": {
        "operationId": "DiagnosesController_reviseDiagnosisAlias",
        "parameters": [
          {
            "name": "diagnosisId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseDiagnosisRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "doctor-decision"
        ]
      }
    },
...
  },
  "info": {
    "title": "DermaHealth Backend API",
    "description": "Clinic workflow backend: identity, scheduling, clinical, workflow, EMR, CRM.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "EncounterResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "parentEncounterId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "appointmentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "standard",
              "emergency",
              "follow_up",
              "remote"
            ]
          },
          "origin": {
            "type": "string",
            "enum": [
              "appointment",
              "walk_in",
              "follow_up_request"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "registered",
              "intake_in_progress",
              "intake_complete",
              "ai_assessed",
              "routed",
              "checked_in",
              "under_doctor_review",
              "awaiting_results",
              "diagnosed",
              "plan_approved",
              "workflow_active",
              "in_progress",
              "results_complete",
              "final_review",
              "discharge_ready",
              "record_signed",
              "closed",
              "post_visit_monitoring",
              "escalated",
              "follow_up_linked"
            ]
          },
          "department": {
            "type": "string"
          },
          "room": {
            "type": "object",
            "nullable": true
          },
          "queueNumber": {
            "type": "object",
            "nullable": true
          },
          "peopleAheadInQueue": {
            "type": "object",
            "nullable": true
          },
          "estimatedWaitMinutes": {
            "type": "object",
            "nullable": true
          },
          "currentDoctorId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "blockingCondition": {
            "type": "object",
            "nullable": true
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "patientId",
          "parentEncounterId",
          "appointmentId",
          "type",
          "origin",
          "status",
          "department",
          "room",
          "queueNumber",
          "peopleAheadInQueue",
          "estimatedWaitMinutes",
          "currentDoctorId",
          "blockingCondition",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateEncounterRequest": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "standard",
              "emergency",
              "follow_up",
              "remote"
            ]
          },
          "origin": {
            "type": "string",
            "enum": [
              "walk_in",
              "follow_up_request"
            ]
          },
          "department": {
            "type": "string"
          },
          "parentEncounterId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "patientId",
          "type",
          "origin",
          "department"
        ]
      },
...
      "DoctorReviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "encounterId": {
            "type": "string",
            "format": "uuid"
          },
          "aiAssessmentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "partial",
              "rejected"
            ]
          },
          "acceptedConditionCode": {
            "type": "object",
            "nullable": true
          },
          "rationale": {
            "type": "object",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "encounterId",
          "aiAssessmentId",
          "doctorId",
          "action",
          "acceptedConditionCode",
          "rationale",
          "reviewedAt"
        ]
      },
      "ReviewAssessmentRequest": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "accepted",
              "partial",
              "rejected"
            ]
          },
          "acceptedConditionCode": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          }
        },
        "required": [
          "action"
        ]
      },
      "DoctorDiagnosisResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "encounterId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "none",
              "provisional",
              "differential",
              "confirmed",
              "revised",
              "signed"
            ]
          },
          "conditionName": {
            "type": "string"
          },
          "conditionCode": {
            "type": "object",
            "nullable": true
          },
          "aiAssessmentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "isAdditionalToAI": {
            "type": "boolean"
          },
          "rationale": {
            "type": "object",
            "nullable": true
          },
          "revisionOfId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "version": {
            "type": "number"
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "encounterId",
          "doctorId",
          "status",
          "conditionName",
          "conditionCode",
          "aiAssessmentId",
          "isAdditionalToAI",
          "rationale",
          "revisionOfId",
          "version",
          "recordedAt"
        ]
      },
      "RecordDiagnosisRequest": {
        "type": "object",
        "properties": {
          "conditionName": {
            "type": "string"
          },
          "conditionCode": {
            "type": "string"
          },
          "aiAssessmentId": {
            "type": "string",
            "format": "uuid"
          },
          "isAdditionalToAI": {
            "type": "boolean"
          },
          "rationale": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "provisional",
              "confirmed"
            ]
          }
        },
        "required": [
          "conditionName",
          "isAdditionalToAI",
          "status"
        ]
      },
      "ClinicalPlanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "encounterId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "diagnosisId": {
            "type": "string",
            "format": "uuid"
          },
          "summary": {
            "type": "string"
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time"
          },
          "autoActivatedWorkflowInstanceId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "Set once the Workflow module auto-activates an instance from this plan."
          }
        },
        "required": [
          "id",
          "encounterId",
          "doctorId",
          "diagnosisId",
          "summary",
          "approvedAt",
          "autoActivatedWorkflowInstanceId"
        ]
      },
      "ApproveClinicalPlanRequest": {
        "type": "object",
        "properties": {
          "diagnosisId": {
            "type": "string",
            "format": "uuid"
          },
          "summary": {
            "type": "string"
          }
        },
        "required": [
          "diagnosisId",
          "summary"
        ]
      },
      "ReviseDiagnosisRequest": {
        "type": "object",
        "properties": {
          "conditionName": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          }
        },
        "required": [
          "conditionName",
          "rationale"
        ]
      },
...
    }
  }
}